sorcery point styles
